WIP SQUASH(?) data type nits
- Prefer size_t for size/length/count specs.
- Prefer u8/u16 over types of unknown size and for consistency.
- Improve stability of sample rate calculation when UART bitrates are
  not a multiple of 9600. (Can this be? Need we care whether it does?
  Think future maintenance, and just be done.)

+++ src/hardware/itech-it8500/protocol.h
@@ -150,7 +150,7 @@ struct dev_context {
 	double max_power;
 	double min_resistance;
 	double max_resistance;
-	uint32_t max_sample_rate_idx;
+	size_t max_sample_rate_idx;
 
 	double voltage;
 	double current;
